+++ src/utils.js
@@ -17,9 +17,6 @@
  */
 const process = require('process')
 const request = require('request-promise-native')
-const redis = require('async-redis')
-
-const logger = require('./logger')
 
 // NOTE: See https://stackoverflow.com/questions/38073527/request-how-to-set-user-agent-for-every-request/38074818#38074818
 const customHeaderRequest = request.defaults({
@@ -29,23 +26,6 @@ const customHeaderRequest = request.defaults({
 })
 
 exports.customHeaderRequest = customHeaderRequest
-
-/**
- * Returns an instance of the database.
- * @function
- * @return RedisClient
- */
-function createClient () {
-  const client = redis.createClient({
-    url: (process.env.REDIS_URL !== undefined) ? process.env.REDIS_URL : 'redis://localhost:6379/1'
-  })
-  client.on('error', async function (err) {
-    logger.error(err)
-  })
-  return client
-}
-
-exports.createClient = createClient
 
 /**
  * Where to find