+++ worker.js
@@ -1,29 +1,45 @@
 const Queue = require('bull')
 const errors = require('request-promise-native/errors')
 
+const Redis = require('ioredis')
 const processSource = require('./src/processors/sources')
 const processEvent = require('./src/processors/events')
-const { createClient, customHeaderRequest, LOCATIONS_URL } = require('./src/utils')
+const { customHeaderRequest, LOCATIONS_URL } = require('./src/utils')
 const { requestToken, createEvents, proposeEvents } = require('./src/endpoint')
 const { currentSource } = require('./src/sources')
 const logger = require('./src/logger')
 
 const REDIS_URL = (process.env.REDIS_URL !== undefined) ? process.env.REDIS_URL : 'redis://localhost:6379/1'
 
+const QUEUE_CLIENT = new Redis(REDIS_URL)
+const QUEUE_SUBSCRIBER = new Redis(REDIS_URL)
+
+const opts = {
+  createClient: function (type) {
+    switch (type) {
+      case 'client':
+        return QUEUE_CLIENT
+      case 'subscriber':
+        return QUEUE_SUBSCRIBER
+      default:
+        return new Redis(REDIS_URL)
+    }
+  }
+}
+
 function handleError (job, err) {
   logger.error(err)
 }
 
 logger.info(`Initializing token queue.`)
-const tokenQueue = new Queue('tokens', REDIS_URL)
+const tokenQueue = new Queue('tokens', opts)
 tokenQueue.on('failed', handleError)
 
 tokenQueue.process(requestToken)
 
 tokenQueue.on('completed', async function (job, result) {
-  const client = createClient()
+  const client = opts.createClient(null)
   await client.set('access_token', result)
-  await client.quit()
   logger.info(`Successfully updated access token`)
 })
 
@@ -31,13 +47,13 @@ tokenQueue.add(null, { repeat: { every: 35000 * 1000 } }) // Repeat every 35000 
 
 // Initialize events queue
 logger.info(`Initializing events queue`)
-const eventsQueue = new Queue('events', REDIS_URL)
+const eventsQueue = new Queue('events', opts)
 eventsQueue.on('failed', handleError)
 
 eventsQueue.process(processEvent)
 
 eventsQueue.on('active', async (job) => {
-  const client = createClient()
+  const client = opts.createClient(null)
   const event = job.data
 
   const count = await client.sismember('processed_events', event.url)
@@ -47,12 +63,11 @@ eventsQueue.on('active', async (job) => {
     await job.discard()
     await job.moveToFailed(new Error(message))
   }
-  await client.quit()
 })
 
 eventsQueue.on('completed', async (job, result) => {
   let event = result
-  const client = createClient()
+  const client = opts.createClient(null)
   await client.sadd('processed_events', event.url)
   try {
     const locations = await customHeaderRequest({
@@ -74,14 +89,12 @@ eventsQueue.on('completed', async (job, result) => {
     } else {
       logger.error(err)
     }
-  } finally {
-    await client.quit()
   }
 })
 
 // Initialize sources queue
 logger.info(`Initializing sources queue.`)
-const sourcesQueue = new Queue('sources', REDIS_URL)
+const sourcesQueue = new Queue('sources', opts)
 sourcesQueue.on('failed', handleError)
 
 sourcesQueue.process(processSource)
@@ -96,13 +109,12 @@ sourcesQueue.on('completed', (job, result) => {
 
 // Initialize current source queue
 logger.info(`Initializing current_source queue`)
-const currentSourceQueue = new Queue('current_source', REDIS_URL)
+const currentSourceQueue = new Queue('current_source', opts)
 currentSourceQueue.on('failed', handleError)
 
 currentSourceQueue.process(async () => {
-  const client = createClient()
+  const client = opts.createClient()
   const source = await currentSource(client)
-  await client.quit()
   return source
 })